Whitelist
This white list script is simple - it allows only white listed people to the server - via a country (in this case: Poland) or serial. You can change the country inside the .lua script. Also if you add a serial to the whitelist of a person from another country without changing the script, it will not let him join because only Polish people can join, but this is just an example - you can use it.