30 FPS
This is a simple script, that allow you to turn on 30 FPS on the map you want. Make sure you change the map name in the .lua file for it to work. Immediately after a map start or restart, the script will do a FPS change to 30, and if you change the map - FPS go back to normal.